Changed base speed, and fixed overflow issue in motion controller

+++ src/motionController.h
@@ -16,16 +16,16 @@ class MotionController {
         float speed;
 
         // Target Motor Encoder Counts
-        int targetMotor1; 
-        int targetMotor2;
+        long targetMotor1; 
+        long targetMotor2;
 
         // Reference to absolute position variables
         float& absoluteX;
         float& absoluteY;
         const volatile SwitchState& last_pressed;
 
-        int motor1Error;
-        int motor2Error;
+        long motor1Error;
+        long motor2Error;
 
         bool moving_completed;
